benchmarks/diffractive_vm: orchestrate using a Snakefile
